+++ src/read.rs
@@ -1,22 +1,22 @@
 use super::*;
 
 /// A read only transaction over the [`EquivalentDB`],
-pub struct ReadTransaction<K, V, S> {
-  db: EquivalentDB<K, V, S>,
-  rtm: Rtm<K, V, HashCm<K, S>, PendingMap<K, V>>,
+pub struct ReadTransaction<K, V, I, C> {
+  db: I,
+  rtm: Rtm<K, V, C, PendingMap<K, V>>,
 }
 
-impl<K, V, S> ReadTransaction<K, V, S> {
+impl<K, V, I, C> ReadTransaction<K, V, I, C> {
   #[inline]
-  pub(super) fn new(db: EquivalentDB<K, V, S>) -> Self {
-    let rtm = db.inner.tm.read();
+  pub(super) fn new(db: I, rtm: Rtm<K, V, C, PendingMap<K, V>>) -> Self {
     Self { db, rtm }
   }
 }
 
-impl<K, V, S> ReadTransaction<K, V, S>
+impl<K, V, I, C> ReadTransaction<K, V, I, C>
 where
   K: Ord,
+  I: Database<K, V>,
 {
   /// Get a value from the database.
   #[inline]
@@ -26,7 +26,7 @@ where
     Q: Ord + ?Sized,
   {
     let version = self.rtm.version();
-    self.db.get(key, version)
+    self.db.as_inner().get(key, version)
   }
 
   /// Returns true if the given key exists in the database.
@@ -37,7 +37,7 @@ where
     Q: Ord + ?Sized,
   {
     let version = self.rtm.version();
-    self.db.contains_key(key, version)
+    self.db.as_inner().contains_key(key, version)
   }
 
   /// Get all the values in different versions for the given key.
@@ -48,21 +48,21 @@ where
     Q: Ord + ?Sized,
   {
     let version = self.rtm.version();
-    self.db.get_all_versions(key, version)
+    self.db.as_inner().get_all_versions(key, version)
   }
 
   /// Returns an iterator over the entries of the database.
   #[inline]
   pub fn iter(&self) -> Iter<'_, K, V> {
     let version = self.rtm.version();
-    self.db.iter(version)
+    self.db.as_inner().iter(version)
   }
 
   /// Returns an iterator over the entries (all versions, including removed one) of the database.
   #[inline]
   pub fn iter_all_versions(&self) -> AllVersionsIter<'_, K, V> {
     let version = self.rtm.version();
-    self.db.iter_all_versions(version)
+    self.db.as_inner().iter_all_versions(version)
   }
 
   /// Returns an iterator over the subset of entries of the database.
@@ -74,7 +74,7 @@ where
     Q: Ord + ?Sized,
   {
     let version = self.rtm.version();
-    self.db.range(range, version)
+    self.db.as_inner().range(range, version)
   }
 
   /// Returns an iterator over the subset of entries (all versions, including removed one) of the database.
@@ -86,13 +86,6 @@ where
     Q: Ord + ?Sized,
   {
     let version = self.rtm.version();
-    self.db.range_all_versions(range, version)
-  }
-}
-
-impl<K, V, S> Clone for ReadTransaction<K, V, S> {
-  #[inline]
-  fn clone(&self) -> Self {
-    Self::new(self.db.clone())
+    self.db.as_inner().range_all_versions(range, version)
   }
 }
